requirements.txt
Django==3.2.9
gunicorn==20.1.0
django-heroku==0.3.1
psycopg2-binary==2.9.1
django-cors-headers==3.10.0
djangorestframework==3.12.4
dj-database-url==0.5.0
django-environ==0.8.1

